requirements.txt
PyNaCl==1.5.0
web3==6.11.3
eth-account==0.9.0
requests==2.31.0
py-solc-x==1.1.1
